third_party/blink/web_tests/platform/generic/virtual/text-antialias/computed-line-height-and-font-size-with-font-size-adjust-expected.txt

font-size-adjust should affect neigther computed line-height nor font-size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS parseFloat(testStyle.getPropertyValue('font-size-adjust')) is within 0.0001 of 0.44
PASS referenceStyle.getPropertyValue('font-size-adjust') is "none"
PASS testStyle.getPropertyValue('font-size') is "200px"
PASS testStyle.getPropertyValue('line-height') is "200px"
PASS containerStyle.getPropertyValue('line-height') is "200px"
PASS successfullyParsed is true

TEST COMPLETE

xx
